Pull getEachFood into domFoods and make SRP
