+++ lib/index.js
@@ -1,5 +1,5 @@
 import { appendMeals, getEachFoodDiary, filterFoods, removeMealFood } from './domIndex'
-import { validateForm } from './domFoods'
+import { validateForm, getEachFood } from './domFoods'
 
 const handleResponse = (response) => {
   return response.json()
@@ -30,35 +30,6 @@ const getAllFoodsValidation = () => {
     .then((response) => handleResponse(response))
     .then((meals) => getAllFoods(meals))
     .catch((error) => console.error({error}))
-}
-
-const getEachFood = (meals,foods) => {
-  let sortedFoods = foods.sort((a,b) => a.id > b.id ? -1 : 1)
-  let foodsInMeals = []
-  meals.forEach((meal) => {
-    meal["foods"].forEach((food) => {
-      foodsInMeals.push(food.id)
-    })
-  })
-  sortedFoods.forEach((food) => {
-      if ($.inArray(food.id, foodsInMeals) === -1) {
-        $('#all-foods').append(`<tr id="${food.id}">
-        <td class="name" id="food-${food.id}">${food.name}</td>
-        <td class="calories" id="cal-${food.id}">${food.calories}</td>
-        <td><button class="delete-btn" aria-label="Delete">
-        Delete
-        </button></td>
-        </tr>`)
-      } else {
-        $('#all-foods').append(`<tr id="${food.id}">
-        <td class="name" id="food-${food.id}">${food.name}</td>
-        <td class="calories" id="cal-${food.id}">${food.calories}</td>
-        <td><button class="delete-btn" aria-label="Delete" disabled>
-        Delete
-        </button></td>
-        </tr>`)
-    }
-  })
 }
 
 const getAllFoodsDiary = () => {
